created last 2 pages, deleted useless coponents, commented home video player and change website background color
