+++ src/App.css
@@ -35,7 +35,7 @@
 }
 /*temporaire*/
 #root{
-  background-color:black;
+  /*background-color:black;*/
 }
