+++ src/components/MusicStyle/styles/MusicStyleComponent.css
@@ -31,7 +31,7 @@
 /*    margin-right: 50px;*/
 }
 #musicWrap{
-    background: #040404;
+    background: #111;
     height: 92vh;
 }
 #musicWrap > .flex{
