+++ src/components/MusicStyleDetails/styles/MusicStyleDetailsComponent.css
@@ -21,7 +21,7 @@
 #header{
     width: 100%;
     height: 8vh;
-    background: #040404;
+    background: #111;
     display: flex;
     justify-content: space-between;
     align-items: center;
@@ -63,7 +63,7 @@
 #anecdote{
     width: 910px;
     height: 80vh;
-    background: #040404;
+    background: #111;
     display: flex;
     align-items: center;
     overflow: hidden;
@@ -158,7 +158,7 @@
     /*background: #a80000;*/
 }
 #wrap{
-    background: #040404;
+    background: #111;
     height: 92vh;
 }
 #wrap > div > section{
@@ -227,7 +227,7 @@
 /*ArtistComponent*/
 #artist{
     margin-top: 50px;
-    background: #040404;
+    background: #111;
     width: 80%;
 }
 #artist > div{
@@ -317,4 +317,16 @@
     left: 0;
     top: 10px;
     position: absolute;
+}
+/* influence */
+#anecdote.influence .video {
+    width: 250px;
+    height: 250px;
+    margin-bottom: 150px;
+}
+#anecdote.influence img {
+    width: 100%;
+}
+#anecdote.influence .text {
+    width: 50%;
 }
